+++ changelog.md
@@ -1,9 +1,20 @@
 # betsol-ng-entity-form changelog
 
+## Version 0.0.2
+(25 Oct 2015)
+
+- Fixed an issue with boolean values
+- Add-ons before for input groups are now supported
+- datetime field type is now also supported
+- Added dependency for [UI Bootstrap](https://angular-ui.github.io/bootstrap/)
+- Added dependency for [Bootstrap-UI/datetime-picker](https://github.com/Gillardo/bootstrap-ui-datetime-picker)
+
+
 ## Version 0.0.1
 (21 Sep 2015)
 
 - Renamed template
+
 
 ## Version 0.0.0
 (21 Sep 2015)